fix vietqr pay

+++ ticketbookingapi/app/utils/fallback_token_manager.py
@@ -24,7 +24,7 @@ class FallbackTokenManager:
         self,
         user_id: int,
         role: str,
-        expires_in_hours: int = 24
+        expires_in_hours: float = 24
     ) -> str:
         """Generate JWT access token"""
         payload = {
